Progress (#5)
* Merge Game/src and Game/include, `Color` docs

* Minor corrections

* Enforce LF line endings for FlatBuffers header generation

* Remove SDL_shadercross as a dependency

* Add shader generator script, provide shadecross Windows executable as tool

* Update shader generation script

* Update scripts

* Update GenerateShaders.py

* Update GenerateShaders.py

* Update GenerateShaders.py

* Update scripts

* Update spatial hash

* Add `GOLDEN_RATIO` math constant

* Update GenerateShaders.py

* Update scripts

* Minor changes

* Math docs

* Add `BoundingSphere` constructor

* Add AABB converter

* Update converters

* Math docs

* Start `Font` class

+++ Source/Math/Objects/Vector3i.cpp
@@ -105,7 +105,7 @@ namespace Silent::Math
             return Vector3(FP_FLOAT(x, shift), FP_FLOAT(y, shift), FP_FLOAT(z, shift));
         }
 
-        return Vector3((float)x, (float)y, (float)z);
+        return Vector3(x, y, z);
     }
 
     const glm::ivec3& Vector3i::ToGlmVec3i() const
